got rid of magic number, and added aiming line, note trajectory speaker
i got rid of the magic number because in the real world, when we shoot against subwoofer, we just barely make it into the top of the speaker. Since our code calculates for shooting into the middle of the speaker, then it will raise the arm a few degrees. I overlooked this when adding the magic number, thinking it was raising the arm for no reason.